updated all tests to use conftest shared file (#110)
